requirements.txt
opencv-python-headless
python-telegram-bot
python-dotenv
